requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.1.0
greenlet==3.1.1
gunicorn==23.0.0
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
typing-extensions==4.12.2
werkzeug==3.1.3
zipp==3.21.0
